requirements.txt
einops
hydra-core
lightning
numpy
omegaconf
pytorch_lightning
scikit_learn
torch
tqdm
kaggle
wandb